nfp_net: pass nfp_net to nn_read[lq]/write[lq] helpers
Since the endian conversions are gone our register
access helpers do nothing but make the code a little
longer.  Convert them to a form which is usual for
most of the Linux drivers - make them take a pointer
to nfp_net and always write ctrl_bar.  This saves us
a few lines of code.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>