nfp_net_common: fix races while starting/waking tx queues
.start_xmit() and tx_complete() can run in parallel so
there is a tiny chance of a race between them where
tx_complete() would already clean the packet which caused
.start_xmit() to stop the queue because .start_xmit() did
not see the cleaning yet.

Note this is extremely unlikely to happen in practice
because we read the HW pointer only at the beginning of
tx_complete() and also we would wake up the queue after
 the next RX on this vector.

Because the restarting code got a little more hairy now
I moved it to complete_tx().

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>