nfp_uio_main: fix builds on Linux v4.3
Linux v4.3 contains changes which will make it possible to use
MSIs on non-PCI buses, in particular:

commit 25a98bd4ff93 ("genirq/MSI: Store 'struct device' instead of
'struct pci_dev' in struct msi_desc")
commit 4a7cc8316705 ("genirq/MSI: Move msi_list from struct pci_dev
to struct device")

We need to compat for that.  For the first one simply pass
pdev down to nfp_uio_msi_mask_irq().  For the second we
need a macro wrapper to keep the code ifdef-free.

While at it correct the types from int32_t to s32.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>