nfp_net_common: go back to the initial encap API
The initial meaning of the descriptor flags was: separate
flags for NVGRE and VXLAN, setting them makes the csum
flags refer to inner headers.

There is no way to supprt outer csum with this API
so that support is dropped completely here.

We also have to parse packet much harder in the tx csum
function.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>