nfp_net_main: use dma_set_mask_and_coherent()
We already use dma_set_mask_and_coherent() to set both masks
in the VF driver.  While at it use the appropriate define
and drop the error message for this super-unlikely error.
Also move the call after pci_request_regions() so that the
ordering is closer to the VF driver.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Acked-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>