nfp_net_common: do not enable scatter support
The driver does not support scatter transfers, do not enable
that feature in firmware.  If FW scatters a frame to two or
more buffers the driver will not handle it correctly.  This
change will save us trouble when the scatter support is added.

Note that setting NETIF_F_SG based on gather support alone is
correct since the stack does not care if we produce scattered
frames or not.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>