nfp_net: make endianness of RSS key more reliable
We kept the RSS key as u32 values while the key is really
a byte array.  It worked fine until we added user space
access to the key: even if we wrote the key in wrong endian
the only source of the values used to be get_random_bytes()
so it didn't really matter if we reshuffled the bytes.
Now that users can set the key to the value they like we
should be careful not to byte-swap it implicitly during copying.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>