nfp_net: perform RSS init only during device initialization
We generated new key on init but the indirection table and
hash options were set during .ndo_open.  Let's bring them
together to init, thanks to this change indirection table
and hash options will not be reset on every ifup.

For completeness write the key on .ndo_open as well (even
though it may not be strictly needed).

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>