nfp_net_common: remove impossible if condition from IRQ handler
We assign RX and TX rings to all r_vectors regardless of
the number of the actual rings in use.  Therefore it is
impossible to receive an interrupt for a vector which
does not have any rings.

While at it move the comment about automasking as it looks
like it was referring to napi_schedule() now instead of
absence of masking.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Acked-by: Jason S. McMullan <jason.mcmullan@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>