nfp_net_common: move link status reading to a separate function
Link status reading code is repeated in the .ndo_open and
the IRQ handler.  Let's move it to a single function.

We have to be carefull to update the nfp_net.link_up whenever
we call netif_carrier_off().

Thanks to this change there will be no more duplicated Link Up
messages during ifup.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>