nfp_net_common: add link status spin lock
We must read link status in .ndo_open to see what link status
was before we could handle interrupts.  If we do it before we
enable the LSC IRQ there is a small window of time when the
status can change and IRQ will still not be ready, yet.

Reading while IRQ is active has the small probability of:

read_from_open:
  sts <- readl(CTRL_BAR_STS);   // reads UP
  <IRQ>
    sts <- readl(CTRL_BAR_STS); // reads DOWN
    netif_carrier_off();
  <EOI>
  netif_carrier_on();

Therefore we need a spinlock to make the status reading, saving
into nfp_net.link_up and netif_carrier_*() atomic.  Note that
nfp_net.link_up must be converted to a standalone field since
bitfields are pretty bad at insuring operations are atomic.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Rolf Neugebauer <rolf.neugebauer@netronome.com>