nfpcore: make hwinfo respect nfp_resource_acquire() error code
hwinfo_fetch_nowait() assumes that if nfp_resource_acquire()
returned an error it must be because the resource was not
found in the resource table and falls back to an old resource
table location to check there.

However, nfp_resource_acquire() may very well fail because
it encountered a real error and we shouldn't try the fallback
at that point, just return the error.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>
Reviewed-by: Dinan Gunawardena <dgunawardena@netronome.com>