nfpcore: make time accounting in hwinfo_fetch() more accurate
Timeout hwinfo_fetch() can be configured by the user
(hwinfo_wait module parameter) therefore we should make sure
that if user said "wait 40 seconds" we are pretty accurate.
msleep() on which we depend for counting time can sleep
for duration which are significantly off the target.  Let's
use jiffies to check if we should still be waiting.

Note that the new code will wait for one iteration more than
strictly required.  This is because we want to avoid situation
where hwinfo_fetch_nowait() blocks for very long time and
"steals" all the time budget.

While at it remove unnecessary variable initialization and
a impossible error condition/nfp_err message.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Jason S. McMullan <jason.mcmullan@netronome.com>
Reviewed-by: Dinan Gunawardena <dgunawardena@netronome.com>