nfpcore: nsp: clean out references to nfp_device
NSP uses struct nfp_device's priv mechanism to store the NSP
resource pointer.  This means that the NSP code has to play
around with the resource lock otherwise it would be locked
from the time the first command is requested to the
moment the nfp_device is closed.

Provide explicit NSP open/close functions.  We can now remove
the resource lock/unlock helpers.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>