merge nfp_netvf.ko into nfp_net.ko
Merge the nfp_net.ko and nfp_netvf.ko modules into one.
CONFIG_NFP_NET_VF can be set on command line to n or y
to request the VF-specific code to be excluded or included
(default) in the build.

Unfortunately kernels < 3.15 can't deal with multiple
device tables in a single module so we need a add a compat
dispatcher for them.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>