nfp_target: move code from a header file to a source
For unclear reasons nfp_target.h is a header file full of static
inline functions.  Move the code to a source file.  This allows
us to drop the underscores from names (we no longer need to mark
functions as "private", they won't be visible at all).  It also
gives a nice decrease in driver size:

-rw-rw-r-- 1 jkicinski jkicinski 338809 Dec 30 22:06 src/nfp_before.ko
-rw-rw-r-- 1 jkicinski jkicinski 337282 Dec 30 22:06 src/nfp_after.ko

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>