nfp_net: spell out firmware names in MODULE_FIRMWARE() macros
In upstream kernel we need to specify the FW names we may use
otherwise mkinitrd may include our driver in initramfs without
them resulting in driver being loaded before it can access the
real file system and the real /lib/firmware.

Upstream we will also use probe deferral if FW image is not found
in initramfs, but deferral still leads to spamming the logs with
errors from many retries so it's best to just inform the kernel
about the FWs we may need.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dinan Gunawardena <dinan.gunawardena@netronome.com>