Makefile: access the .revision file from Kbuild
We used to load the contents of .revision file from the Makefile.
This doesn't work for DKMS, which always invokes the Kbuild
directly.  We could pass the version directly as a variable
but DKMS doesn't make that very easy, so just load .revision
from Kbuild.  Note that Kbuild is evaluated from within src/
so the actual path to the file is ../.revision.  We also allow
placing the revision file directly in the sources, because
DKMS builds may not control ../.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Simon Horman <simon.horman@netronome.com>