tools: add a debug script for dumping NSP logs
Add a script for dumping all NSP logs on a system.  These are currently
rarely provided and they can be crucial for reasoning.  The script will
simply wrap ethtool -w if netdev is present, but in many cases netdev
isn't there, so the script will try to load the driver with nfp_dev_cpp=1
and use the nfp-arm tool.

There are no parameters, switches and options to the script, people
should just execute it and it will "do the right thing (tm)".

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>