compat: improve the Ubuntu mainline build detection
Ubuntu's mainline build kernels have a fairly useless
UTS_UBUNTU_RELEASE_ABI define which can't be trusted to
be a valid number.  We used to check if there are any
characters in it (like "rc"), this doesn't catch invalid
octal constants like 041805.  Parse the define out and
use [[ $VAL -ge 0 ]] to determine if it can be treated as
an integer.

Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Dirk van der Merwe <dirk.vandermerwe@netronome.com>