nfp: compat: fix error in tc indirect block backport
A recent patch added compat code to deal with the function name change for
indirect block registering in the kernel. This was done by moving to the
new name and adding compat helpers that call the old function on older
kernels. Unfortunately, the functions contain a typedef that is not
present in kernels older than 5.0 and so cause compilation issues on such
systems.

Add kernel version macros to ensure compat code can compile on kernels
older than 5.0.

Signed-off-by: John Hurley <john.hurley@netronome.com>
Acked-by: Jakub Kicinski <jakub.kicinski@netronome.com>