nfp: compat: add comparison functions of RHEL_RELEASE for CentOS-stream
As a rolling-release Linux distro, the update of CentOS-stream may cause
some compatibility issues which can only be filtered through RHEL_RELEASE.
It may be helpful to extract the RHEL_RELEASE and add some comparison
functions of it.

The format of RHEL_RELEASE is temporarily defined as xxx.x.x which can be
modified in future updates.

Signed-off-by: Fei Qin <fei.qin@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>