workflow: dkms: packaging: add DKMS packaging workflow
A 'ci-coordinator' workflow is added to call individual
workflows to run based on the branch name. This workflow
will be used to call the DKMS packaging workflow followed
by testing workflows afterwards.

A 'package-dkms' workflow is added, which is called by
the 'ci-coordinator' workflow, to build RPM and DEB DKMS
packages.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>