packaging: dkms: rpm: add DKMS rpm build script
This patch adds support to build DKMS RPM packages of the
nfp driver. An RPM spec.in file is added which is used
when building the package. From DKMS version 3, support
was removed to build RPM packages directly from DKMS.
This is why this patch introduces a manual build process
instead of using DKMS packaging tools.

- The spec.in file is a modified version of the one
  included in previous versions of DKMS. The file was
  updated to include the correct package name, version,
  company information and other modifications required to
  get a working build.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>