packaging: update '.revision' file location
The '.revision' file is used to show the driver version in ethtool. When
building the driver using DKMS, the location 'src/../.revision' does not
work correctly, because DKMS uses a symlink to the source directory. One
directory down does not include the '.revision' file, which means the
driver version in ethtool is not displayed. Moving the '.revision' file
into the source directory ensures it is correctly used by DKMS.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>