nfpcore: cpp: fix up quirk_nfp6000 detection
This changes affects the out-of-tree driver only, as it forms part
of the extra code bits that enables cpp-access to userspace tools.

With the recent "nfp: support Corigine PCIE vendor ID" change
the #ifndef variable changed to use the new define, but now this
is not fulfilling the original purpose of effectively being a
slightly smarter Kernel Version check. Revert this back to using the
original define.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>