workflow: fixup compile tests on older kernels
"make rpm-pkg" for older kernels does not end up including all the
requirements for building out-of-tree kernel modules. Or at least
not in the same way newer kernels does. Update these jobs to just
use the compiled kernel directly instead of going through the steps
of creating a rpm and unpacking that again. This means we loose the
benefit of caching from previous builds, but the runtime is still
around 5 minutes, which is acceptable.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Danie du Toit <danie.dutoit@corigine.com>