compat: fix kernel v3.8 compatibility
There is a small bug in kernel v3.8 which causes the warning:
    <drv>/src/nfp_net_compat.c: In function ‘devl_assert_locked’:
    <drv>/src/nfp_net_compat.c:222:17: error: unused variable ‘pf’ [-Werror=unused-variable]
      struct nfp_pf *pf = devlink_priv(devlink);
                     ^
    cc1: all warnings being treated as errors

This is because the macro "lockdep_assert_held(&pf->lock);" gets
expanded to nothing when CONFIG_LOCKDEP is not set. This was fixed
in v3.9 by: Commit: 5cd3f5affad2 ("lockdep: Silence warning if CONFIG_LOCKDEP
isn't set").

Add a specific check for kernel v3.8 to get around this.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>