compat: updates for kernel 4.18.0-348 on CentOS 8.5.2111
TCA_PEDIT_KEY_EX_CMD_SET is defined in tc_pedit.h
which is included in the condition as below:
 #if (defined(COMPAT__HAVE_METADATA_IP_TUNNEL) ||  \
     LINUX_VERSION_CODE >= KERNEL_VERSION(5, 1, 0))

but the TCA_PEDIT_KEY_EX_CMD_SET is used in function
which is in #else branch as below:
 #if VER_NON_RHEL_LT(5, 1) || VER_RHEL_LT(8, 1)
 ...
 #else
 ...
 #endif

When compiling on CentOS 8.5.2111(kernel 4.18.0-348),
it is possible for tc_pedit.h to not be included, because
the linux version(4.18.0) is less than 5.1.0. But
the function which uses the TCA_PEDIT_KEY_EX_CMD_SET
will be compiled, because the RHEL version(8.5.2111)
is greater than 8.1.0

This problem have mostly been hidden away by the definition of
COMPAT__HAVE_METADATA_IP_TUNNEL in the Kbuild file, but
it's good to have this fixed for consistency.

Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>