packaging: rpm: add warning message if kernel-devel is not installed
The installed version of kernel-devel is not necessarily the correct
version for the current kernel version. This is because if you install
the kernel-devel package without specifying the version, the latest
version available will be installed, not the correct version for the
current running kernel.

This behaviour causes the DKMS package to be installed, but the driver
is not built. A warning message is added to at least warn the user that
the driver will not be built, and how to install the correct package.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>