workflows: ci-coordinator: modify triggers
The workflow is modified to only trigger on pushes for branch names
starting with the specified prefixes. These changes reflect the
intended use cases, which are:

- test-feature:
  Triggers on push, runs build and tests.
- build-feature:
  Triggers on push, runs build.
- release-version:
  Triggers on push, runs build.

This is updated since it was found that it happens surprisingly
often that developers create a branch with -test in the name to try things
out, without actually meaning to trigger any tests. Limit the triggers
to starting prefixes only to alleviate this problem.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: James Hershaw <james.hershaw@corigine.com>