workflow: ci-coordinator: add an all job status check
This patch includes a notification that will be sent out
and ran only if this is a workflow running on public-main
or a release branch as well as a prerelease branch. It makes
use of the Zabbix sender utility to send this notification
to Zabbix where an email will be sent out.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Danie du Toit <danie.dutoit@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>