nfp: allow receiving all multicast packets in kernel<=v3.15
The function `__dev_mc_sync` used to sync device’s multicast
list is not implemented in kernel v3.15 or earlier versions,
so we add compatibility code in OOT driver so that multicast
filter feature is not available in those earlier versions.

However multicast filter is enabled by driver once it’s
supported in FW, so all the multicast packets are dropped in
the earlier kernel versions. We need disable this feature in
earlier kernel versions to fix this issue.

Fixes(o-o-t): 32cb6b21091e ("nfp: add support for multicast filter")
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>