nfp: fix incorrect compat code for compiling IPSec feature
IPSec feature implementation depends on `xa_array`, which is introduced
in v4.20. So v4.20 is the first kernel version for nfp to support IPSec
currently. But some APIs are adjusted in v5.1, like `xa_alloc` which is
also used in our IPSec code, which makes it compiled error in v4.20 and
v5.0.

Now modify the compat code to fix compiling error, consequently it delays
to v5.1 to start to support IPSec now. I think it's OK since some main
Linux distributions like Centos8 have already backported the latest
`xa_array` implementations, so the change won't affect that much.

Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>