nfp: add compat codes for Kylin V10 SP3
Support is required to use driver on Kylin V10 SP3. However, some
compatibility problems are found. Add some compat codes for Kylin
V10 SP3.

The kernel source of Kylin is unavailable. Lots of test cases
failed in tests.flower with the in-tree driver. Thus, we disable
the flower build against Kylin. This can be revisited in future
should flower support be needed in the future.

Signed-off-by: Fei Qin <fei.qin@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>