workflow: checkout_corigine: update checkout actions
This patch updates the checkout actions to make use of the
`checkout_corigine` github action in ci-libs.

I have also added a clean up step. This required the `findutils`
package to be added to the `Install pre-requisite packages` step
since it does not get installed by default on the docker.

Signed-off-by: Zanelle Vorster <zanelle.vorster@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Reviewed-by: Dano Teixeira <dano.teixeira@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>