workflows: add back missing token in some checkout steps
The 'token: ${{ secrets.GHUB_TOKEN }}' line was accidentally wrongly
removed in a few places where it shouldn't have been.

The token is not required if the repo you are checking out is the same
repo as the one that triggered the workflow, but if you are checking out
a different private repo it is needed. So in this case - checking out
the nfp-drv-kmods repo does not require it, but the steps checking out
ci-libs does need it.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Reviewed-by: Zanelle Vorster <zanelle.vorster@corigine.com>