tools: update Corigine IRQ affinity script
This update allows the target CPU cores to be specified as either
a range "0-4" or a comma-separated list "0,1,4", or a combination
of these formats. The script reports which CPU cores are local to
the target netdev, but does not restrict IRQ binding to these cores,
giving the user full control. The SMP affinity and XPS masks are
set from the list of cores specified. A critical distinction is
that this script assigns IRQs to all available cores specified
(non-unique assignment).

The functionality was necessitated by the need to reserve some
local NUMA cores for user space applications such as iperf.

If no core specification argument is provided, all local NUMA
nodes are used for affinity binding.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>