packaging: include tools in packaging
Changes in this patch include:

A profile.sh script is added to the tools/ directory to be installed
along with the driver. This will add '/opt/netronome/drv' to the PATH
when sourced, which is where any '<nfp-drv-kmods>/tools' scripts
are installed.

Modifications to the process for packaging the driver as a .rpm/.deb
package to install:
* 'set_irq_affinity.sh' as '/opt/netronome/drv/nfp_set_irq_affinity'
* 'tools/profile.sh' as '/etc/profile.d/nfp_drv_kmods_dkms_profile.sh'

The same modification was also added to the 'make install' target
of the normal Makefile when compiling from source.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>