tools: add nfp_troubleshoot_gather script
This imports a python script that can be used to log troubleshooting
information for the NFP's on the system. The script logs several bits of
information using standard linux tools and built-in python methods to
save information for debugging.

It also contains a check to see if the nfp Board Support (BSP) package
is installed. If it is and the driver is loaded with the 'nfp_dev_cpp=1'
flag it will also log various bits of lower level firmware/card
information.

By default the BSP log path will dump general information about all the
NFP's in the system. If an interface name is specified through the
'--interface' argument it will also dump more specific information
applicable to the specific card the netdev is associated with.

The default logging directory is:
    '/var/tmp/nfp_troubleshoot_gather_<date>'

A temporary directory will be created here during data collection, after
which it will be tarred up into a .tgz file, and the temporary directory
will be deleted.

The log directory can be overridden with '--logdir' argument.

It is highly recommended to run the script as superuser, though some
basic information will still be logged if not.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>