workflows: packaging: update ownership within docker
A recent update seems to have broken the old workaround of adding '--add
safe.directory' to the global git config. Not sure what component was
updated, it is assumed to be the git version that is in use.

Fix this by changing the ownership of the working directory to that of
root. The directory is cleaned up afterwards, so this should not lead to
any long term permission issues.

As part of this also move to using the shared actions, which means that
the ci-libs repo does not need to be checked out out anymore.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>