workflow: packaging: override the __python3 rpm macro
With the addition of the 'nfp_troubleshoot_gather.py' script in the
driver-dkms package the required dependencies changed. This makes sense,
as python3 is now required. However, the automatic dependency generation
adds a "Requires:" section for '/usr/libexec/platform-python', which is
not available on some .rpm based systems. Fix this by overriding the
default %{__python3} macro value to point to '/usr/bin/python3' instead.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>