workflows: package-dkms: folder per month compatibility
The outputs in the cloud storage folders becomes more and more.
To make it more user friendly, folders now get created for every month
for releases and pre-releases. The format of the folder name is "YY.MM".

The folders for wip builds are updated to be more user friendly by
creating separate directories for each user. The folder name is based on
the actors GitHub user name.

The updates in this patch include:
  - After the dkms binaries are built, the files get uploaded to a
    separate folder per month for releases and pre-releases. The format
    of the folder name is "YY.MM".
  - In the case of interim builds, the destination directory remains
    unchanged.
  - In the case of work in progress builds, the binaries get uploaded to
    the temporary nfp-drv-kmods directory under the username of the
    GitHub user that triggered the workflow.

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>