workflows: ci-coordinator: use tags for releases and prereleases
Currently workflows are triggered on the creation of a
release/prerelease branch and when a pull request is created into a
release/prerelease branch. However, when a release/prerelease branch
gets created, it does not necessarily contain the commits for the
release. After the branch is created, additional commits are added to
the branch, making the automated testing not very useful.

The release model has been changed to start with the creation of a
release branch for the month, then commits can be added to the
release branch. Once all the necessary commits are added to the
release branch, a tag is created for the prerelease. This tag is
fixed for the prerelease and it should not be moved. If bugs need to be
fixed or additional commits need to be added for some reason, the
commits can be added to the release branch. Once the new batch of
commits are added, a new tag can be created for the new prerelease. Once
everything is stable, a release tag can be created at the commit
with the last prerelease tag. If additional fixes are needed, they can
be added to the release branch and the tagging methodology can
continue.

In terms of the automation, the workflows will be triggered in the
following cases:
- Creation and push of the release branch: only linters
- Pull requests into the release branch: drv-regressions
- Creation of a prerelease tag: package_dkms & install_dkms_packages
- Creation of a release tag: package_dkms & install_dkms_packages

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>