build: add support for tags in describe-head.sh
Currently, describe head only looks at the branch of the current commit
and not the TAG. With the recent switch to using tags for releases, this
has become a problem when trying to get the versions. Thus,
functionality is added that allows the user to specify the ref using an
environment variable.

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>