tools: troubleshoot: check for both 'sosreport' and 'sos'
Turns out that the sosreport tool is sometimes called 'sos', in which
case the command is 'sos report', and sometimes it is called
'sosreport', which is then used as-is. Introduce a helper to check for
both and return the appropriate command to use. Update the 'sos_data'
helper to take in this command as input argument.

As part of this update also cleanup the 'sos_data()' function,
it was doing unnecessary 'echo' and 'cat' commands, instead
of just using the string directly.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>