tools: troubleshoot: add check for required tools
Some basic tools are required for the troubleshoot script to work,
like 'tar', which is not guaranteed to be installed on a system.

Add a check which can be used to for the presence of these required
tools. For now this is only tar, but should we want to 'force' more
of the other tools in use by the script to be present we can easily
add it to the list.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>