workflow: ci-test-interop: fix azure search pattern
When triggering the interoperability tests from a wip branch,
the azcopy_list action cannot find the packages because it is looking in
the wrong directory. The package is uploaded to a directory with a
specific date in the directory name thus the workflow is modified to
look for the package in the AZURE_PATH and it looks for the specific
extension using the binary_type.

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Dano Teixeira <dano.teixeira@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>