Makefile: update module installation directory
The directory in which the nfp modules will be installed by the
modules_install target is dependent on the kernel version, unless
otherwise specified.

This directory is now specified as INSTALL_MOD_DIR=updates, the change
to install to the updates directory instead of extra is because the
extra directory is not on the depmod search path for all systems,
whereas updates is.

The paths for uninstalling these modules are also changed to remove the
files from the updates directory.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>