workflow: ci-test-interop: clean yum/dnf cache
Use distro-sync to synchronize the installed packages on the system with
the latest available packages in the repositories. This command ensures
that the system is in sync with the latest state of the distribution.

Sometimes, the yum cache may contain outdated information. Thus the
cache is updated and cleaned before the installation of packages.

Signed-off-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>