workflows: add .tgz packaging to create-packages workflow
This updates the create-packages.yaml workflow to specifically specify
a bin_type as part of the input matrix, instead of just relying on the
OS-type to determine the package type. Use this new field to add in the
.tgz creation step.

Also update a few more places from dkms to 'distro' or 'package'.
The current upload folder on Azure is called `nfp-drv-dkms`, this will
be a more significant change and can impact quite a lot of other
things, so leaving that in place for now.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>