workflows: package_rpm: add kylinOS v10_SP3 docker
Add KylinOS to the list of distro options. At this point add in only
SP3, as it seems like SP2 and SP1 has some GPG key issues which will
need more investigation.

As part of this patch also add in a step to print the user inputs for
easier referencing when trying to repeat the same build.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>