workflows: package-rpm: define _topdir macro for building RPMs
Define the _topdir macro when building the kernel driver static binaries
to avoid uploading unintended files, which were built in the default
rpmbuild location, to Azure.

This improves upon the previous functionality, which would build in the
default location, and inspect the output of the build command for the
newly created files and move them to a known location.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>