workflows: ci-linting: Disable-on-push for `public-main`
Many commits to the `public-main` branch are being cherry-picked and
pushed which causes the linter to fail. This is because the linter will
lint older commits and not just the current changes since it thinks
they are unique commits. Therefore, the linter is disabled for pushes
from that branch.

Additionally, the `python_version` is set to `2` for the `code_check`
linter.

Signed-off-by: Zanelle Botha <zanelle.botha@corigine.com>
Reviewed-by: Marthinus Bekker <marthinus.bekker@corigine.com>
Reviewed-by: Joshua Van Straaten <joshua.vanstraaten@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>