nfp: compat: avoid redefinition of function
Ensure that the function "netif_is_rxfh_configured()" is only defined
once by using appropriate pre-processor macros. This change is required
to compile for the following commit in the kernel:

Commit 49675f5bdf9a ("net: remove IFF_* re-definition") which removes
the macro definitions from "include/linux/netdevice.h" that were being
used to determine whether or not the function should be defined in the
"nfp_net_compat.h" file.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>