workflows: package-rpm: Set up and run in own docker
Remove github's docker engine and replace it by setting up a
docker environment that is run in the background. Stop and remove
the docker environment at the end of the workflow.

Run the commands in the created container. In order for this to work
all the jobs needed to be switched to use docker, since the
local build action now uses the docker command as its shell, which
breaks non-dockerized use.

Any Node16 issues are solved by running commands in the
container.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Shivasen Govender <shivasen.govender@corigine.com>