compat: Fix bad compat code for RHEL 8 and BC Linux 8
The initial logic for the compat code could result in a scenario where
the first branch of the compat conditional was true, and so a function
prototype for "nfp_flower_setup_indr_block_cb()" was never declared,
even if the condition "VER_RHEL_GE(8, 3)" was also true.

This in turn would result in the function definition being non-static,
which causes the build to fail due to a lack of a function prototype.

Fixes: 81f0dd4a5a82 ("nfp: enable Flower build against RHEL 8 and BC Linux 8 kernels")
Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>