workflows: build: Check for running docker before stopping
While the steps that stops the docker containers have a
'continue-on-error: true' setting it does still lead to quite a lot of
noise on the summary page, saying all jobs exited with a non-zero exit
code. Fix this by actually doing docker checks before stopping and
removing the instance.

The package-rpm jobs seems to have been missing the docker cleanup step
at the end, also add this in.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>