workflows: build-local: fix incorrect 'upload_dir' key usage
The 'upload_dir' variable does not exist anymore as part of the
release matrix. Replace it with 'release.name' instead.

At the same time the substitution in the command to get the version was
using single "{}" brackets, which meant that this always complained
about 'bad substitution'. This is now fixed by using the correct
"${{ }}" scheme.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>