Revert "workflows: update OpenEuler mirrors"
This reverts commit 0f611187b84e8f022be9d2edc20ed164cb2efb86.

It seems like the newly added http://mirror.accum.se/mirror is more
unstable - the URL is up but the package information is incorrect.

The old URL looks to be back up, and apart from the few days of
downtime the other day it has been stable most of the time.

Revert the previous commit to switch back to the original
'https://mirrors.dotsrc.org/' mirror.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>