compat: align defines after Linux 6.15 release
Linux 6.16-rc1 is about to be released. Bump the module version,
we are working on 6.17 now.

Also redefine 'LINUX_VERSION_CODE' for 6.16. It is expected that this commit
is only required for a very short duration. This check is needed
because the merging of ipsec-next into net-next broke some compat code.
This happened during the initial merge window for 6.16, before
6.16-rc1, but after 6.15 release. As soon as v6.16-rc1 is fully
released the 'LINUX_VERSION_CODE' redefine can be adjusted for 6.17.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>